requirements/requirements-testing.txt
coverage == 5.1
# Multienvironment testing
tox
# django_coverage_plugin
sqlparse
sqlalchemy
factory-boy==2.12.0

# mock is a library for testing in Python.
# It allows you to replace parts of your system under test with mock objects and
# make assertions about how they have been used.
mock==3.0.5
